+++ cashmanagerapp/lib/models/article_model.dart
