+++ cashmanagerapp/lib/pages/articlesPages/list_articles.dart
@@ -1,4 +1,4 @@
-import 'package:cashmanagerapp/models/articlemodel.dart';
+import 'package:cashmanagerapp/models/article_model.dart';
 import 'package:cashmanagerapp/pages/detailPages/detail.dart';
 import 'package:cashmanagerapp/services/article_service.dart';
 import 'package:flutter/material.dart';
