+++ cashmanagerapp/lib/pages/detailPages/detail.dart
@@ -1,4 +1,4 @@
-import 'package:cashmanagerapp/models/articlemodel.dart';
+import 'package:cashmanagerapp/models/article_model.dart';
 import 'package:cashmanagerapp/services/article_service.dart';
 import 'package:flutter/material.dart';
 import 'package:cashmanagerapp/widgets/button_like.dart';
